# docs/openapi.yaml
openapi: 3.0.1
info:
  title: L0 Auth
  description: Authorization Service
  license:
    name: MIT
    url: https://github.com/tiki/l0-auth/blob/main/LICENSE
  version: 0.0.6
servers:
- url: https://auth.l0.mytiki.com
paths:
...
  /api/latest/oauth/token:
    post:
      tags:
      - AUTH
      summary: Token Grant
      description: Issue authorization token. Use password grant for OTP flow.
      operationId: l0-auth-oauth-token-post_1_1
      parameters:
      - name: grant_type
        in: query
        description: "(password, refresh, urn:ietf:params:oauth:grant-type:jwt-bearer)"
        required: true
        schema:
          type: string
      - name: scope
        in: query
        required: false
        schema:
          type: string
      - name: username
        in: query
        required: true
        schema:
          type: string
      - name: password
        in: query
        required: true
        schema:
          type: string
      - name: audience
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
      - name: refresh_token
        in: query
        required: true
        schema:
          type: string
      - name: assertion
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2AccessTokenResponse'
  /api/latest/oauth/revoke:
    post:
      tags:
      - AUTH
      summary: Revoke Token
      description: 'Revoke a refresh token. '
      operationId: l0-auth-oauth-revoke-post
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    OAuth2AccessToken:
      type: object
      properties:
        tokenValue:
          type: string
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        tokenType:
          $ref: '#/components/schemas/TokenType'
        scopes:
          uniqueItems: true
          type: array
          items:
            type: string
    OAuth2AccessTokenResponse:
      type: object
      properties:
        accessToken:
          $ref: '#/components/schemas/OAuth2AccessToken'
        refreshToken:
          $ref: '#/components/schemas/OAuth2RefreshToken'
        additionalParameters:
          type: object
          additionalProperties:
            type: object
    OAuth2RefreshToken:
      type: object
      properties:
        tokenValue:
          type: string
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    TokenType:
      type: object
      properties:
        value:
          type: string
...